Set Lambda timeout to 55 seconds
We run every 2 minutes, but /could/ reasonably run every 1 minute, so we
don't want to go to 60 seconds. I've noticed that, on occasion, we start
getting occasional timeouts (likely as logging and metrics start having
increased timeouts in their service). So we can just add a little extra
time to compensate.

* [6 Weeks](https://brwyatt-public-files.s3.amazonaws.com/SmartThings_Logging/Commits/Lambda_Timeouts_6W.png)
* [1 Week](https://brwyatt-public-files.s3.amazonaws.com/SmartThings_Logging/Commits/Lambda_Timeouts_1W.png)